--- .github/CODEOWNERS
@@ -1,4 +1,0 @@
-
-# CodeHierarchyNodeParser
-llama_index/llama-index-integrations/readers/llama-index-readers-file/llama_index/readers/file/code @ryanpeach
-llama_index/llama-index-integrations/readers/llama-index-readers-file/tests/code @ryanpeach
