+++ llama-index-core/llama_index/core/command_line/mappings.json
@@ -279,8 +279,8 @@
   "LLMQuestionGenerator": "llama_index.core.question_gen",
   "SubQuestionOutputParser": "llama_index.core.question_gen",
   "ReaderConfig": "llama_index.core.readers",
-  "CodeHierarchyNodeParser": "llama_index.readers.file",
-  "CodeHierarchyKeywordQueryEngine": "llama_index.readers.file",
+  "CodeHierarchyNodeParser": "llama_index.packs.code_hierarchy",
+  "CodeHierarchyKeywordQueryEngine": "llama_index.packs.code_hierarchy",
   "StringIterableReader": "llama_index.core.readers",
   "ResponseMode": "llama_index.core.response_synthesizers",
   "BaseSynthesizer": "llama_index.core.response_synthesizers",
