+++ llama-index-integrations/readers/llama-index-readers-file/llama_index/readers/file/__init__.py
@@ -16,10 +16,6 @@ from llama_index.readers.file.tabular.base import PandasCSVReader
 from llama_index.readers.file.unstructured.base import UnstructuredReader
 from llama_index.readers.file.video_audio.base import VideoAudioReader
 from llama_index.readers.file.xml.base import XMLReader
-from llama_index.readers.file.code import (
-    CodeHierarchyNodeParser,
-    CodeHierarchyKeywordQueryEngine,
-)
 
 __all__ = [
     "DocxReader",
@@ -42,6 +38,4 @@ __all__ = [
     "ImageTabularChartReader",
     "XMLReader",
     "PagedCSVReader",
-    "CodeHierarchyNodeParser",
-    "CodeHierarchyKeywordQueryEngine",
 ]
