--- llama-index-integrations/readers/llama-index-readers-file/llama_index/readers/file/code/__init__.py
@@ -1,4 +1,0 @@
-from llama_index.readers.file.code.code_hierarchy import CodeHierarchyNodeParser
-from llama_index.readers.file.code.index import CodeHierarchyKeywordQueryEngine
-
-__all__ = ["CodeHierarchyNodeParser", "CodeHierarchyKeywordQueryEngine"]
