+++ llama-index-packs/llama-index-packs-code-hierarchy/llama_index/packs/code_hierarchy/BUILD
@@ -1,0 +1,1 @@
+python_sources()
